new changes

+++ app/assets/javascripts/controllers/my_restaurant.js
@@ -1,21 +1,23 @@
 RestaurantsBa.MyRestaurantController = Ember.ObjectController.extend({
-   actions: {
+   	actions: {
 	   	delete: function(restaurant_id) {
-	   		alert(restaurant_id);
-	   		var self = this;
-	   		$.ajax({
-	   			url: "api/restaurants/" + restaurant_id,
-	   			type: "DELETE",
-	   			success: function(data) {
-	   				self.transitionToRoute("restaurants");
-	   			}
-	   		});
+	   		if (window.confirm("Are you sure you want to delete your restaurant?")) {
+		   		
+		   		var self = this;
+		   		$.ajax({
+		   			url: "api/restaurants/" + restaurant_id,
+		   			type: "DELETE",
+		   			success: function(data) {
+		   				self.transitionToRoute("restaurants");
+		   			}
+		   		});
+		   	}
 	   	},
 	   		edit: function() {
 	   		
 	   		var self = this;
 	   		self.transitionToRoute('edit_restaurant');
 	   	}
-   }
+   	}
 
 });
